cdk/requirements.txt
attrs==20.3.0
aws-cdk.assets==1.72.0
aws-cdk.aws-apigateway==1.72.0
aws-cdk.aws-applicationautoscaling==1.72.0
aws-cdk.aws-autoscaling==1.72.0
aws-cdk.aws-autoscaling-common==1.72.0
aws-cdk.aws-autoscaling-hooktargets==1.72.0
aws-cdk.aws-batch==1.72.0
aws-cdk.aws-certificatemanager==1.72.0
aws-cdk.aws-cloudformation==1.72.0
aws-cdk.aws-cloudfront==1.72.0
aws-cdk.aws-cloudwatch==1.72.0
aws-cdk.aws-codebuild==1.72.0
aws-cdk.aws-codecommit==1.72.0
aws-cdk.aws-codedeploy==1.72.0
aws-cdk.aws-codeguruprofiler==1.72.0
aws-cdk.aws-codepipeline==1.72.0
aws-cdk.aws-codepipeline-actions==1.72.0
aws-cdk.aws-cognito==1.72.0
aws-cdk.aws-ec2==1.72.0
aws-cdk.aws-ecr==1.72.0
aws-cdk.aws-ecr-assets==1.72.0
aws-cdk.aws-ecs==1.72.0
aws-cdk.aws-efs==1.72.0
aws-cdk.aws-eks==1.72.0
aws-cdk.aws-elasticache==1.72.0
aws-cdk.aws-elasticloadbalancing==1.72.0
aws-cdk.aws-elasticloadbalancingv2==1.72.0
aws-cdk.aws-events==1.72.0
aws-cdk.aws-events-targets==1.72.0
aws-cdk.aws-iam==1.72.0
aws-cdk.aws-kinesis==1.72.0
aws-cdk.aws-kinesisfirehose==1.72.0
aws-cdk.aws-kms==1.72.0
aws-cdk.aws-lambda==1.72.0
aws-cdk.aws-logs==1.72.0
aws-cdk.aws-route53==1.72.0
aws-cdk.aws-route53-targets==1.72.0
aws-cdk.aws-rds==1.72.0
aws-cdk.aws-s3==1.72.0
aws-cdk.aws-s3-assets==1.72.0
aws-cdk.aws-sam==1.72.0
aws-cdk.aws-secretsmanager==1.72.0
aws-cdk.aws-servicecatalog==1.72.0
aws-cdk.aws-servicediscovery==1.72.0
aws-cdk.aws-sns==1.72.0
aws-cdk.aws-sns-subscriptions==1.72.0
aws-cdk.aws-sqs==1.72.0
aws-cdk.aws-ssm==1.72.0
aws-cdk.aws-stepfunctions==1.72.0
aws-cdk.cloud-assembly-schema==1.72.0
aws-cdk.core==1.72.0
aws-cdk.custom-resources==1.72.0
aws-cdk.cx-api==1.72.0
aws-cdk.region-info==1.72.0
cattrs==1.1.1
constructs==3.2.20
jsii==1.14.1
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.3
